consumer/pyproject.toml
[project]
name = "consumer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "duckdb>=1.2.2",
    "fastavro>=1.10.0",
    "kafka-python>=2.2.6",
    "phonenumbers>=9.0.5",
    "pytest>=8.3.5",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]
